+++ docs/release-notes.md
@@ -1,5 +1,16 @@
 # Release Notes
 
+## 0.3.0 - 2024-xx-xx
+
+### Enhancements
+
+- Change measurement time function to async.get_event_loop().time()
+- Change MySQL drive from PyMySQL to mysql-connector-python
+- Restrict PR Merge Targets
+- Asynchronous logging
+- Connection class Abstraction
+- Initial Implementation of OpenAPI
+
 ## 0.2.0 - 2024-08-17
 
 ### Features